# config/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Domain Graph
# during dev, should point to your local machine
host: localhost:10010
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /query:
    x-swagger-router-controller: query
    get:
      description: Query graph by domain
      operationId: queryByDomain
      parameters:
        - name: domain
          in: query
          description: Exact domain name to query for
          required: true
          type: string
        - name: depth
          in: query
          description: Depth to traverse in graph
          required: true
          type: number
        - name: clientId
          in: query
          description: Client ID of socket.io connection; must pass to subscribe to changes
          required: true
          type: string
      responses:
        "200":
          description: Subset of the graph starting at domain
          schema:
            $ref: "#/definitions/Node"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
      tags:
        - query
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  GraphResponse:
    properties:
      nodes:
        type: array
        items:
          $ref: "#/definitions/Node"
      links:
        type: array
        items:
          $ref: "#/definitions/Link"
  Node:
    properties:
      id:
        type: integer
        format: int32
        description: Unique ID generated by Neo4j
      domain:
        type: string
        description: Domain Name
      ip:
        type: string
        description: IP Address
  Link:
    properties:
      source:
        type: integer
        format: int32
        description: Neo4j ID of one node
      target:
        type: integer
        format: int32
        description: Neo4j ID of other node
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string